Moved Downloader classes to their own file and updated imports
